added Delte and updated functionality

+++ src/services/apiCabins.js
@@ -1,51 +1,60 @@
-import supabase from "./supabase";
-import { supabaseUrl } from "./supabase";
-// A fucntion to get cabins
+import supabase, { supabaseUrl } from "./supabase";
 
-export async function createCabin(newCabin) {
+export async function getCabins() {
+  const { data, error } = await supabase.from("cabins").select("*");
+
+  if (error) {
+    console.error(error);
+    throw new Error("Cabins could not be loaded");
+  }
+
+  return data;
+}
+
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabins/public/${imageName}`;
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
 
-  console.log(imagePath);
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error("Cabin could not be created");
   }
 
-  //2. Upload Image
-  console.log(newCabin.image);
-  // const avatarFile = event.target.files[0];
-  const { error: storageError } = await supabase.storage
-    .from("cabins")
-    .upload(`public/${imageName}`, newCabin.image);
+  // 2. Upload image
+  if (hasImagePath) return data;
 
-  //3. Delete cabin if Image uplaod was uncessesfull
+  const { error: storageError } = await supabase.storage
+    .from("cabin-images")
+    .upload(imageName, newCabin.image);
+
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.error("cabin image could not uploaded Cabin was not created ");
+    console.error(storageError);
+    throw new Error(
+      "Cabin image could not be uploaded and the cabin was not created"
+    );
   }
 
   return data;
-}
-
-export async function getCabins() {
-  try {
-    const { data, error } = await supabase.from("cabins").select("*");
-
-    return data;
-  } catch (error) {
-    throw new Error("Cabins could not be loaded");
-  }
 }
 
 export async function deleteCabin(id) {
@@ -55,4 +64,6 @@ export async function deleteCabin(id) {
     console.error(error);
     throw new Error("Cabin could not be deleted");
   }
+
+  return data;
 }
